don't caount resources of pending pods and terminated init containers
